Stop lingering daemon container.
In some cases, a lingering daemon container would be present when an
attempt was made to connect with `--docker`, causing docker errors like:
"The container name NN is already in use by container XX ...". When this
happens, the connect logic will now give the container some time to stop
(the normal case is that it just hasn't yet) and then call `docker stop
NN` to stop it, before retrying the start.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>