Let traffic-agent route traffic to pod-IP instead of 127.0.0.1.
Routing traffic to localhost will effectively bypass any injected
sidecar that modifies the iptables for incoming traffic. This commit
mitigates that problem by routing to the pod-IP instead.

The fix currently applies when the intercepted service's `targetPort`
is symbolic. More work is needed in our init-container to also enable
this for numeric ports.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>