Fix race condition when obtaining routes.
Under certain conditions, the routing table may reference interfaces
that no longer exists. This caused failures in our CI-tests such as
"Unable to get interface at index NN". THe situation occurred when a
session was terminated just before an attempt was made to obtain the
table.

This commit adds a retry and a delay to the process of obtaining a
consistent table. If no interface is found at an index, the table
is re-fetched and rescanned.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>