Remove the RootDaemon annotation altogether. It's never used.
The root-daemon is always needed when a session is established, and with
the helm-command moved to CLI, a session is always established when
the user-daemon is used. So this commit removes the `RootDaemon`
annotation.

This behavior is also more consistent with a containerized daemon which
never uses a root daemon.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>