Restore TLS annotation value expansion.
Users of Telepresence <2.9.0 that make use of the pod template
TLS annotations are unable to upgrade because the annotation values
are no longer exposed to environment expansion. It used to be possible
to name a secret "$_TEL_AGENT_NAME-tls-cert" and get that expanded.

This commit restores this behavior in the traffic-manager by using a
fake environment that contains the `_TEL_AGENT_NAME` env-var. It also
adds a `AGENT_NAME` env-var so that the old name (which is very special
and no longer in use) can be replaced.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>