Make the Telepresence client docker image configurable
The docker image used when running a Telepresence intercept in docker
mode can now be configured using the setting `images.clientImage` and
will default first to the value of the environment variable
`TELEPRESENCE_CLIENT_IMAGE`, and then to the value preset by the
telepresence binary. This configuration setting is primarily intended
for testing purposes.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>