Generalize the kubeconfig patcher and gRPC service.
This commit moves some code from the docker package to the patcher
package and to the k8s_config.go file, so that it's clear that the code
isn't docker specific.

The `authenticator` application is also removed in favor of using a
`telepresence kubeauth` subcommand. This, so that the installation using
one single binary will still work when kubeconfig exec patching is
necessary.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Kelvin Piroddi <kpiroddi@yieldstreet.com>