Improve outbound traffic performance.
The root-daemon now communicates directly with the traffic-manager
instead of routing all outbound traffic through the user-daemon. The
root-daemon uses a patched kubeconfig where `exec<` configurations to
obtain credentials are dispatched to the user-daemon. This to ensure
that all authentication plugins will execute in user-space. The old
behavior of routing everything through the user-daemon can be restored
by setting `cluster.connectFromUserDaemon` to `true` in `config.yml`

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Kelvin Piroddi <kpiroddi@yieldstreet.com>