Add direct port-forwarding from TUN device to destination.
Adds a `GetPortForwardPod` function to the traffic-manager RPC so that
a client can ask the traffic-manager what pod and port to forward to
when a connect request is made on a TCP port on an IP address served by
the TUN device.

Kubernetes does not allow port-forward of IP-addresses. It must have
a resource type and a port. If the resource is a service, then the
forward will ultimately go to a pod that the service selects.

The `GetPortForwardPod` will resolve TCP <ip>:<port> requests to both
services and pods and return the final pod's name and namespace together
with the port to forward to. This means that the resolution from service
to pod is made in the traffic-manager which is faster than having the
client do it using additional calls to the kubelet. The traffic-manager
uses its watcher backed states that keep track of services and pods
to make the process of finding the pod as quick as possible.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>