Add WatchAgentPods service to traffic-manager.
This service makes it possible for clients to watch for changes in
intercepted traffic-agents. The watcher will only notify the client
about changes that are relevant the client's session. It is triggered
both when by agent changes and when agents change their intercept
status with respect to the client session.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>