Use traffic-agent port-forwards for outbound and intercepted traffic.
The telepresence TUN-device is now capable of establishing direct
port-forwards to a traffic-agent in the connected namespace. That
port-forward is then used for all outbound traffic to the device, and
also for all traffic that arrives from intercepted workloads. Getting
rid of the extra hop via the traffic-manager improves performance and
reduces the load on the traffic-manager. The feature can only be used if
the client has Kubernetes port-forward permissions to the connected
namespace. It can be disabled by setting `cluster.agentPortForward` to
`false` in `config.yml`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>