Introduce intercept disposition type REMOVED and modified timestamp.
We want to track personal intercepts that have been REMOVED and let them
linger for some time period while the agent responds with an error.
This commit adds the disposition type REMOVED so that the existing
WatchIntercepts subscription can be used to track that state.

This commit also separates the actual remove from running finalizers
in the traffic-manager state, so that finalizers can run when the user
leaves the intercept and the actual remove can happen later on.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>