Untangle nested locking causing deadlock during root daemon disconnect.
On rare occasions, the VIF device would end up in a function that
required a read lock of the `Endpoint` that the gvisor stack uses and
that would be impossible due to the stack's Wait method calling the
device's Wait method, which in turn waited for a `sync.WaitGroup`
that never got done, because it should have been released by the
caller of the first function. Classic deadlock.

This commit ensures that the `stack.Wait` never calls on the device's
`Wait` and instead calls the device's `Endpoint.Wait` directly by
renaming the device's `Wait` to `WaitForDevice`. That function is then
called explicitly in serial with `stack.Wait`. This untangles the
potential deadlock.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>