Keep telepresence-agents configmap in sync with pod states.
An intercept attempt that resulted in a timeout due to failure of
injecting the traffic-agent left the `telepresence-agents` configmap
with a state that indicated that an agent had been added.

The next time an intercept was made, presumably after fixing what caused
the failure, the traffic-manager would detect the existing entry in the
configmap and leave it untouched and hence, not trigger the necessary
rollout needed for the injection to take place.

This commit ensures that the configmap entry is deleted when there is
a failure to inject the traffic-agent. This keeps the configmap in sync
with the actual state of the pods.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>