Repair broken clusterInfoEqual function.
The `clusterInfoEqual` function used when determining the necessity of
notifying clients about changes in the cluster configuration had
apparently been forgotten when adding new information that needed to
be tracked dynamically. Specifically, it did not compare (and hence no
notifications were sent for changes of) `alsoProxySubnets`,
`neverProxySubnets`, `allowConflictingSubnets`, `includeSuffixes`, or
`excludeSuffixes`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>