Ensure that connector.log exists before starting docker daemon.
Our CI-tests on linux-arm64 uses a rootless docker installation. After
a successful runs with that container, it is no longer possible to start
a non-containerized user daemon and no user daemon is visible in the
collected logs. Presumably, this is because the connect.log have
permissions that doesn't allow it to be overwritten och even copied.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>