Provide agent name and tag as separate values from Helm chart
The `AGENT_IMAGE` was a concatenation of the agent's name and tag. This
is now changed so that the env instead contains an `AGENT_IMAGE_NAME`
and `AGENT_INAGE_TAG`. The `AGENT_IMAGE` is removed.

Also, a new env `REGISTRY` is added, where the registry of the traffic-
manager image is provided. The `AGENT_REGISTRY` is no longer required
and will default to `REGISTRY` if not set.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>