Make server CIDR discovery use IPv6 when services are IPv6.
The CIDR discovery makes an attempt to create a service using in IP
that is known to be outside the service CIDR range. This provokes an
error where the actual service CIDR is mentioned, and can be extracted.

This error is different when the IP-protocol is incorrect. This commit
ensures that the protocol is the same as the IP for the agent-injector
service.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>