Run sshfs in slave mode when podIP is IPv6
One would assume that adding square-brackets around the IP would be
sufficient, but sshfs cannot handle that. Internally, it'll strip it off
and concatenate with the port number using a colon, with an "Address
family for hostname not supported" error as the result.

The solution was to open the connection to the server ourselves, and
then use a bidirectional pipe to connect stdin/stdout of sshfs running
with `-o slave`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>