Ensure that detection of kind/minikube uses IPv6 when possible.
The logic that swaps the hosts address to the cluster for the address
used from within a docker container didn't take into account that the a
host address of `[::1]:<port>` must be swapped with a corresponding
IPv6 address in the container.

For kind clusters, this commit also replaces the hostname used before
with an IP, which saves time that was previously spent on failing DNS
lookups once the cluster was connected.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>