Restore DNS dropSuffixes and use that for the tel2-search too.
The `dropSuffixes` is a list of suffixes to drop before processing the
query. Historically, this was used by the overriding resolver to drop
suffixes added from `/etc/resolv.conf`, because those suffixes are
intended for the fallback only. This list can also contain the
"tel2-search" domain because it should be treated the same way.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>