Disable root-daemon's metriton reporting when SCOUT_DISABLE=1
The root daemon starts without environments being passed on from the
caller (passing them requires special permissions), so it is not
possible to disable the metriton reporter by just setting
`SCOUT_DISABLE=1` in the environment.

This commit adds a boolean flag `--disable-metriton` to the daemon. The
CLI passes this flag when `SCOUT_DISABLE=1` variable is set, so that the
daemon can set it too.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>