Fix correct generation of files under /etc/resolver for darwin systems.
This commit simplifies how files are generated under `/etc/resolver`, so
that one file per DNS-route is generate without a `search`, and one
file per DNS-search is generated with a `search`.

If the `search` is suffixed with an existing file generated for a
`route`, then that file will receive the `search`.

Files are no longer suffixed with `.local` because there's really no
reason to do that.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>