Fix kubeconfig exec authentication with context names containing colon
The logic added to allow the root daemon to connect directly to the
cluster using the user daemon as a proxy for exec type authentication in
the kube-config, didn't take into account that a context name sometimes
contains the colon ":" character. That character cannot be used in
filenames on windows because it is the drive letter separator.

Closes #3488

Signed-off-by: Thomas Hallgren <thomas@datawire.io>