Restore ability to run the telepresence CLI in a docker container.
The improvements made to be able to run the telepresence daemon in
docker using `telepresence connect --docker` made it impossible to run
both the CLI and the daemon in docker. This commit fixes that and also
ensures that the user- and root-daemons are merged in this scenario when
the container runs as root.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>