Deactivate agents that are about to die.
When a change is made to an entry in the `telepresence-agents`
comfigmap, then the pods of the targeted workflow will be restarted.
It is then of some importance that they no longer participate as
agents for intercepts.

This commit ensures that such pods become inactive, which means that
they will no longer be able to modify intercepts with a call to
`ReviewIntercept`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>