Fix potential TLS-configuration corruption in injector's tls listener
The TLS configuration passed to the `tls.Server` was not properly
isolated with respect to concurrency. It had a one element slice of
certificates that could get its element replaced by another go-routine.

This commit ensures that each `tls.Server` gets its own configuration
and that the contained certificate is passed by value.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>