Add ValidateCreateAgent function.
This function is meant to validate that it is OK to create a new entry
in the `telepresence-agents` configmap, with all that it then triggers
in the form of rollout and agent injection. The enterprise version uses
this to assert that we can create an Istio Sidecar CRD for the agent.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>