Remove large number of requests when traffic-manager starts up.
When the traffic-manager starts, it creates a watcher for services.
This watcher will receive all services in from all namespaces that are
managed. For a global manager, that means all services in the cluster.

Prior to this commit, each service that was received, would result in
three list queries, one for all deployments that matched the service's
selector, then a similar query for replicasets, and another for
stateful sets. So for a cluster with 1000 services, there would be
3000 list queries using selector match at startup.

These queries were part of an inefficient way of finding out which
agents that might be affected by the services, and are now replaced
by an algorithm that instead looks at the `telepresence-agents`
configmap initially, and then if necessary (for entries that doesn't
match by service UID), it performs a query for the workload whose kind
is known by the config. This means that there will be at maximum one
query per intercepted workload.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>