Don't rely on context name when deciding if kind is used.
The code relied on the context name starting with "kind-", and yes
all contexts created by kind have that name, but the user is free to
rename it or to create other contexts using the same connection
properties.

This commit changes the check to instead look for a loopback service
address.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>