Ensure that agents are considered busy when used as proxy-via.
An agent connection established from the root daemon was only considered
busy when it was intercepted, or when it was the last connection
remaining. This meant that when a proxyVia arrived and more connections
was present, it wasn't considered busy, and therefore ignored.

This commit ensures that agents used for proxy-via are considered busy
at all times.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>