Retry downloads on Windows
One of the most common failures in our CI is when windows downloads
the wintun.dll. Other common failures are downloads of the winfsp.msi
and sshfs-win.msi, and the choco install of make and zip.

This commit wraps all of those in `nick-fields/retry`, in hope of
improving the build stability.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>