Let traffic-manager's PrepareIntercept reuse mutators agent Map
There's no need for code in the traffic-manager to create and manipulate
the `telepresence-agents` configmap. All of that can now be done by the
mutator's `Map`, which incidentally also uses a cache.

This commit removes all such code in favor of reusing the `Map`, so that
potential race-conditions are removed and the number of API accesses
is minimized.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>