Include "svc" as a top-level domain in the DNS resolver.
It's not uncommon that use-cases involving Kafka or other middleware use
FQNs that end with "svc". The core-DNS resolver in Kubernetes can
resolve such names. With this bugfix, the Telepresence DNS resolver will
also be able to resolve them, and thereby remove the need to add ".svc"
to the include-suffix list

Closes #2814

Signed-off-by: Thomas Hallgren <thomas@datawire.io>