Docker aliases deprecation caused failure to detect Kind cluster.
The logic for detecting if a cluster is a local Kind cluster, and
therefore needs some special attention when using `telepresence connect
--docker`, relied on the presence of `Aliases` in the Docker network
that a Kind cluster sets up. In Docker versions from 26 and up, this
value is no longer used, but the corresponding info can instead be found
in the new `DNSNames` field.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>