Removal of ability import legacy artifact into Helm.
A helm install would make attempts to find manually installed artifacts
and make them managed by Helm by adding the necessary labels and
annotations. This was important when the Helm chart was first introduced
but is far less so today, and this legacy import was therefore removed.
Maintaining it properly would mean adding regression tests and to fix
pending issues

Signed-off-by: Thomas Hallgren <thomas@datawire.io>