Do not reset connection consumption metrics on every call to remain.
The time spent by a connection must be accumulated, and time spent when
the connection is dormant (no remain calls arrive from the client) must
not be counted.

This commit ensures that all remain-calls add the proper time and that
sequences of more than 4 missed remain calls are counted as one. This
means when a user closes the laptop lid, the connection will be
considered dormant.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>