FailedScheduling events mentioning node availability considered fatal
The traffic-manager considers some events as fatal when waiting for
traffic-agent to arrive after an injection has been initiated. This
logic would trigger on events like "Warning FailedScheduling 0/63 nodes
are available" although those events indicate a recoverable condition
and kill the wait. This is now fixed so that the events are logged but
the wait continues.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>