Warn when connecting to an Enterprise traffic-manager from an OSS client
The difference between the OSS and the Enterprise offering is not well
understood, and OSS users often install a traffic-manager using the Helm
chart published at getambassador.io. This Helm chart installs an
enterprise traffic-manager, which is probably not what the user would
expect. Telepresence will now warn when an OSS client connects to an
enterprise traffic-manager and suggest switching to an enterprise
client, or use `telepresence helm install` to install an OSS
traffic-manager.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>