Introduce DNS fallback on Windows.
Telepresence belongs to a virtual network device with excellent network
performance, and is therefore often rated with a low metric and placed
first in the list of DNS servers.

Utilities using `nslookup` or `node.resolveXXX` calls fail when
Telepresence is connected, because those methods will use the first
resolver in the list and then give up when it fails.

This commit introduces a `network.defaultDNSWithFallback` option on
Windows. It will cause the DNS-resolver to fall back to the resolver
that was first in the list prior to when Telepresence establishes a
connection. The option is default `true` since it is believed to give
the best experience but can be set to false to get the old behavior.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>