Prevent all updates to telepresence-agents configmap from the injector.
Updates to the telepresence-agents configmap may well trigger a roll-out
of the corresponding workload. That roll-out will trigger new calls to
the injector. If the API is fast enough, this will result in a loop that
is similar to a recursion. The result is a large number of created pods
that eventually must be deleted.

This commit removes the update of the configmap from the injector so
that this loop is prevented. The only reason for this update to take
place was to cater for the annotation
`telepresence.getambassador.io/inject-traffic-agent: enabled`. A
subsequent commit will add watchers that deal with that particular case.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>