Add workload watchers and event handlers to deal with inject annotation.
Adds watchers for the `ReplicaSets` and `StatefulSets` to the existing
`SharedInformer` so that all types of workloads are watched.

Also add the needed event handlers that take care of the special
`telepresence.getambassador.io/inject-traffic-agent` annotation, and
ensure that the traffic-agent is injected or removed in accordance with
its setting.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>