Add the ability to use spinners when building and running docker images.
Also makes the message shown when an intercept handler container becomes
ready configurable. This message may vary depending on what the
container is doing. If it started a debugger, then the message should
state that we're waiting for a debugger frontend to attach. If not, the
message should probably be "type <ctrl>-C to exit...".

Signed-off-by: Thomas Hallgren <thomas@datawire.io>