Detect minikube network when connecting with --docker
A `telepresence connect --docker` failed when attempting to connect to a
minikube that uses a docker driver because the containerized daemon did
not have access to the "minikube" docker network. Telepresence will now
detect an attempt to connect to that network and attach it to the daemon
container as needed.

Signed-off-by: Thomas Hallgren <thomas@tada.se>