Use ghcr.io for images and a Minikube or Colima cluster.
This PR uses a Minikube or Colima cluster instead of kubeception. This
removes the need for a special proprietary kubeception token when
running the integration tests, and also opens up for running tests that
require Docker on all platforms.

The PR also changes how images are built and used. Instead of pushing
them a registry in the cluster, the images are pushed to ghcr.io, and
then removed when the workflow ends.

Integration tests on windows are disabled because there's no way to run
a local cluster without nested virtualization. In essence:
 - WSL1 doesn't work because it doesn't provide systemd, and systemd
   requires a real kernel.
 - WSL2 is unable to run because virtualization is unavailable.

Signed-off-by: Thomas Hallgren <thomas@tada.se>
Signed-off-by: Thomas Hallgren <thomas@datawire.io>