Fix backward incompatibility regression.
The introduction of argo-rollouts introduced a call to the new
traffic-manager function `GetKnownWorkloadKinds` from the user client,
and a client connecting to older traffic-managers that don't have this
method would therefore fail with an error.

This commit checks if the returned error has code `Unimplemented`, and
if so, assumes the default set of workload kinds.

Signed-off-by: Thomas Hallgren <thomas@tada.se>