Fix problem with root daemon not starting when sudo timeout is zero.
The root daemon refused to start when `sudo` was configured with a
`timestamp_timeout=0`. This was due to logic that first requested root
privileges using a sudo call, and then relied on that these privileges
were cached, so that a subsequent call using `--non-interactive` was
guaranteed to succeed. This logic will now instead do one single sudo
call, and rely solely on sudo to print an informative prompt and start
the daemon in the background.

Signed-off-by: Thomas Hallgren <thomas@tada.se>