Change traffic-agent's target port to podIP instead of localhost.
Using the IP of the pod means that an application now has a choice to
either bind to that IP or to localhost. Internally, it also meant that
we could implement a safer routing from the traffic-agent to the app-
container when numeric ports were used.

In detail, there's a huge difference between requests that the traffic-
agent performs on behalf of a client that wants to connect to the app,
and the forwarding it does to the app during times when no intercepts
are active. The former must be routed back to the agent, so that
potential intercepts are served correctly, whereas the latter must not
be routed back, because that would result in an endless loop.

Closes #3473

Signed-off-by: Thomas Hallgren <thomas@tada.se>