Make usage data collection configurable using an extension point
Replace the function `scout.NewReporter` with a variable so that the
actual function can be configured at runtime. The default setting for
the variable is a function that renders all usage data collection as
no-ops.

The `Report` interfaced used is left untouched so that the metriton
implementation that is now removed can be re-added in the enterprise
client.

Signed-off-by: Thomas Hallgren <thomas@tada.se>