Enable intercepts of workloads that have no service.
Telepresence is now capable of intercepting workloads that have no
associated service. The intercept will then target container port
instead of a service port. The new behavior is enabled by adding a
`telepresence.getambassador.io/inject-container-ports` annotation where
the value is a comma separated list of port identifiers consisting of
either the name or the port number of a container-port, optionally
suffixed with `/TCP` (the default) or `/UDP`.

Closes #1647

Signed-off-by: Thomas Hallgren <thomas@tada.se>