Auto generate docs/releaseNotes.md from CHANGELOG.yml
Adds a new `relnotesgen` tool that generates a reasonable markdown
formatted file from the yaml described in our CHANGELOG.yml. This tool
is then used in the `make generate` target to generated the
`docs/releaseNotes.md`.

The generation means that the release notes are readable already in this
repository. It also means that a lot of logic can be removed from the
telepresence.io website.

Signed-off-by: Thomas Hallgren <thomas@tada.se>