Add an option to specify target container when intercepting.
An intercept ultimately targets a specific port within a container. The
port is usually determined by examining the relationship between the
service's `targetPort` and the container's `containerPort`.

In certain scenarios, the container owning the intercepted port differs
from the container the intercept targets. This container's sole purpose
is to route traffic from the service to the intended container, often
using a direct localhost connection.

This commit introduces a `--container <name>` option to the intercept
command. While this option doesn't influence the port selection, it
guarantees that the environment variables and mounts propagated to the
client originate from the specified container. Additionally, if the
`--replace` option is used, it ensures that this container is replaced.

Signed-off-by: Thomas Hallgren <thomas@tada.se>