Some workloads missing in the telepresence list output
Version 2.20.0 introduced a regression in the `telepresence list`
command, resulting in the omission of all workloads that were owned by
another workload. The correct behavior is to just omit those workloads
that are owned by the supported workload kinds `Deployment`,
`ReplicaSet`, `StatefulSet`, and `Rollout`. Furthermore, the `Rollout`
kind must only be considered supported when the Argo Rollouts feature is
enabled in the traffic-manager.

Signed-off-by: Thomas Hallgren <thomas@tada.se>