Request install-ID prior to launching the daemon.
With Metriton removed, there's no one requesting the install-ID prior to
launching the user daemon. If that daemon is launched in a container, it
will have the config ina read-only filesystem and therefore fail to
store a generated install-ID.

This commit ensures that the ID is requested (and created if it doesn't
exist) before starting the containerized daemon.

Signed-off-by: Thomas Hallgren <thomas@tada.se>