Crash in traffic-manager configured with agentInjector.enabled=false
A traffic-manager that was installed with the Helm value
`agentInjector.enabled=false` crashed when a client used the commands
`telepresence version` or `telepresence status`. Those commands would
call a method on the traffic-manager that panicked if no traffic-agent
was present. This method will now instead return the standard
`Unavailable` error code, which is expected by the caller.

Signed-off-by: Thomas Hallgren <thomas@tada.se>