Add the ability to use a name for the target host.
Knowing the IP of the local container that acts as the handler service
for an intercept, replace, or wiretap is not possible until the
container has been started.

This commit introduces the concept of target host synthetic IP. A named
target host will be replaced by a random synthesized IPv6 IP when the
engagement is initialized. Later, when the traffic-agent container dials
this IP, the original name will be found and resolved into an IP, thus
deferring the name resolution until an actual request is made to the
engaged pod.

This allows the engagement to be created before the creation of the
handler service, and removes the need for the user to use IP addresses.

Signed-off-by: Thomas Hallgren <thomas@tada.se>