Regression causing "unexpected slice size" with older traffic-managers.
Older traffic-managers have a different way of reporting the
service-subnet. The new way, using a list of subnets reused a proto
slice in the GRPC message that was expected to be empty, but older
traffic-managers will pass the IP of the kube-dns here. It cannot be
parsed as a list of subnets. A check that remedies this mismatch was
inserted.

Signed-off-by: Thomas Hallgren <thomas@tada.se>