Telepresence fails to start the root daemon on Windows
The telepresence CLI starts a user daemon and a root daemon. The latter
is started using administrator privileges. On a Windows box, this means
that the root daemon runs using a different user account (typically
"Administrator") unless the current user can run processes with elevated
privileges. The socket used for communication with the root daemon was
assumed to reside in `%USERPROFILE%\AppData\Local\telepresence` and was
therefore not found by the CLI and the user daemon. The location will
henceforth always be based on the `%USERDATA` of the CLI user.

Signed-off-by: Thomas Hallgren <thomas@tada.se>