Engagement disconnects after certain amount of time
The configuration parameter `connectionTTL`, controlling how long a
client could be completely idle before the traffic-manager or
traffic-agent would consider it dead and disconnect (default 24 hours),
had no effect. Instead, an engagement would disconnect after 2 hours
(the default gRPC `keepAlive.Time` duration). The default of 24 hours is
now reinstated.

The Helm value `client.connectionTTL` was moved to `grpc.connectionTTL`
because it is a server configuration. The old value will still work, but
it is deprecated and will be removed eventually.

Closes #3887
Closes #3856

Signed-off-by: Thomas Hallgren <thomas@tada.se>