gather-logs produces no cluster-side logs when connected with --docker
The `telepresence gather-logs` command did not include cluster-side logs
when connected using `--docker`, because it tried to store such logs in
a temporary directory created by the CLI. The directory was not mounted
in the daemon container. This is now changed so that the temporary
directory is created under the users cache directory, which is
guaranteed to be mounted on the container.

Signed-off-by: Thomas Hallgren <thomas@tada.se>