The default port for the mutating webhook is now 8443. It used to be 443
Port numbers below 1000 are reserved for privileged processes and are
often restricted by firewalls. Consequently, the default port for the
mutating webhook was changed from 443 to 8443. You can override this
default port using the agentInjector.webhook.port value in the Helm
Chart. This change is particularly significant for clusters using
Telepresence, where firewall rules limit the admission webhook's access
to worker nodes, such as in an Amazon EKS cluster.

Signed-off-by: Thomas Hallgren <thomas@tada.se>