A connect --docker fails when the cluster's address is on localhost
Telepresence would fail to connect to a cluster from within a container
if the kubeconfig pointed to a port-forwarded localhost because that
localhost is not reachable from within the container.

This situation is now detected so that the address used from within the
container has "localhost" replaced with "host.docker.internal", or an
alternative alias configured by the user using the new
`docker.hostGateway` parameter in the client configuration.

Signed-off-by: Thomas Hallgren <thomas@tada.se>