More efficient DNS handling in the traffic-manager
Telepresence will no longer send DNS queries for A and AAAA records to
the traffic-manager. Instead, it will send a single query for the name
and then derive the record type from the type of IP address (IPv4 or
IPv6) in the response. This reduces the number of DNS queries sent to
the cluster's DNS server and makes the behavior more consistent with the
`net.LookupNetIP` function in the Go standard library, which the
traffic-manager ultimately uses.

The lookups will now be performed exclusively by the traffic-manager,
never by the traffic-agent. This means that traffic-agents with special
DNS configurations might stop working. If this is a problem, the old
behavior can be restored by setting the `client.dns.useComplexLookup`
parameter in the Helm chart or the `dns.useComplexLookup` parameter in
the client configuration file.

Signed-off-by: Thomas Hallgren <thomas@tada.se>