Introduce an interceptController that controls the intercept lifecycle
Each intercept (or wiretap) must have context associated with it that
gets canceled when the intercept becomes inactive. This commit
introduces an `interceptController` that enforces this. The controller
is canceled when the intercept goes out of scope.

An `interceptControllerMap` type is also introduced which handles common
methods on a map of `interfaceControllers`.

Signed-off-by: Thomas Hallgren <thomas@tada.se>